# src/main/resources/api.yaml
swagger: '2.0'
info:
  title: Sample API
  description: API description in Markdown.
  version: 1.0.0
host: api.example.com
basePath: '/v1/star-wars'
schemes:
  - https
paths:
  /watch-order:
    get:
      summary: Returns a list of star wars episodes.
      operationId: watchOrder
      produces:
        - application/json
      parameters:
        - name: sort
          in: query
          description: Sorting episodes order by
          type: string
      responses:
        200:
          description: OK
          schema:
            type: 'array'
            items:
              $ref: '#/definitions/EpisodeDetails'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Notifications'

  /favorite/{imdbId}:
    get:
      summary: get favorite of a star wars episode
      operationId: getFavoriteEpisode
      parameters:
        - name: imdbId
          in: path
          description: movie id in imdb
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            type: boolean
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Notifications'
    post:
      summary: Add a favorite star wars episode
      operationId: addFavoriteEpisode
      parameters:
        - name: imdbId
          in: path
          description: movie id in imdb
          required: true
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Notifications'
    delete:
      summary: Remove a favorite star wars episode
      operationId: removeFavoriteEpisode
      parameters:
        - name: imdbId
          in: path
          description: movie id in imdb
          required: true
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Notifications'

  /favorite/all:
    get:
      summary: Get all favorite star wars episodes
      operationId: getFavoriteEpisodeAll
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              type: string
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/Notifications'

definitions:
    EpisodeDetails:
      type: object
      properties:
        imdbId:
          type: string
        title:
          type: string
        released:
          type: string
        directors:
          type: string
        actors:
           type: string
        plot:
          type: string
    Notifications:
      type: object
      properties:
        errorCode:
          type: string
        errorMessage:
          type: string